- fully backport DoFlash func - fix bug where BuildingSizeLimit check will skip tunnel check entirely (by deadreaperz) - fix out of bound array acces by mouse cursors (by deadreaperz) - fix unnessesarry functioncall used on SuperClass::AI function (by deadreaperz) - fix bug where prereq doest check for alternate (by deadreaperz) - fix parser bug on GenericPrequisite (by deadreaperz) - fix value type of AICaptureWoundedMark (by deadreaperz) - fix AuxPowerDrain condition  (by deadreaperz) - fix wrong type used when checking HasFactory (by deadreaperz) - fix missing Dusguise parent data (by deadreaperz) - fix Promote Play SpotLight elite code path was using veteran settings (by deadreaperz) - remove duplicate condition on JumpjetLocomotionClass_State3_UpdateSensors hook  (by deadreaperz) - fix broken condition of ExtraThreat (by deadreaperz) - revert change from MIN_VAL to MAX_VAl for FootClass::TryGrindning (by deadreaperz) - disable dead code on function CheckTargetInRange - change some enum definition to enum class to avoid poisoning the project

+++ YRpp/RulesClass.h
@@ -603,7 +603,7 @@ public:
 	TypeList<int >		 AICaptureLowPower;
 	TypeList<int >		 AICaptureLowMoney;
 	int				 AICaptureLowMoneyMark;
-	int				 AICaptureWoundedMark;
+	float				 AICaptureWoundedMark;
 	TypeList<int >		 AISuperDefenseProbability;
 	int				 AISuperDefenseFrames;
 	float				 AISuperDefenseDistance;
